requirements.txt
Flask==3.0.2
Jinja2==3.1.3
PyMuPDF==1.18.18
nltk==3.6.2
spacy==3.7.5
pdfminer.six==20201018
